+++ aidigger/libdigger/src/join_pic_detect.h
@@ -5,7 +5,7 @@ char* WEIGHTS_FILE;
 char* COCONAME;
 
 void print_bytes(unsigned char* bytes, int len, char* name);
-int join_pic_detect(int rand_seed, const char** picNames,unsigned char* result, void* network_ptr, unsigned long thread);
+int join_pic_detect(long rand_seed, const char** picNames,unsigned char* result, void* network_ptr, unsigned long thread);
 void print_hello_join_pic_detect();
 void* initNetwork(char *cfgfile,char *weightfile);
 void enter_to_continue();
